ai-songwriter-prosthesis/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain_community
langgraph
pydantic
PyYAML
pronouncing
serpapi

# --- Tracing & Observability ---
arize-phoenix
openinference-instrumentation-langchain
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi